requirements.txt
pillow
torch
numpy
transformers
opencv-python
timm
accelerate>=0.26.0